ACS-7543: Moving some Auth related code to common-authentication module.
