+++ live-ingester/src/main/java/org/alfresco/hxi_connector/live_ingester/adapters/auth/AuthenticationService.java
@@ -2,7 +2,7 @@
  * #%L
  * Alfresco HX Insight Connector
  * %%
- * Copyright (C) 2024 Alfresco Software Limited
+ * Copyright (C) 2023 - 2024 Alfresco Software Limited
  * %%
  * This file is part of the Alfresco software.
  * If the software was purchased under a paid Alfresco license, the terms of
@@ -25,36 +25,27 @@
  */
 package org.alfresco.hxi_connector.live_ingester.adapters.auth;
 
-import static org.apache.hc.core5.http.HttpHeaders.AUTHORIZATION;
+import static org.alfresco.hxi_connector.common.adapters.auth.AuthSupport.CLIENT_REGISTRATION_ID;
 
 import java.time.Duration;
-import java.util.Map;
-import java.util.Set;
+import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 import jakarta.annotation.PostConstruct;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.camel.CamelContext;
-import org.apache.camel.Exchange;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.boot.autoconfigure.security.oauth2.client.OAuth2ClientProperties;
 import org.springframework.scheduling.TaskScheduler;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
-import org.springframework.security.oauth2.client.authentication.OAuth2LoginAuthenticationToken;
-import org.springframework.security.oauth2.core.OAuth2AccessToken;
-import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.security.oauth2.core.user.OAuth2UserAuthority;
 import org.springframework.security.scheduling.DelegatingSecurityContextTaskScheduler;
 import org.springframework.stereotype.Service;
 
+import org.alfresco.hxi_connector.common.adapters.auth.AuthSupport;
 import org.alfresco.hxi_connector.live_ingester.adapters.config.IntegrationProperties;
-import org.alfresco.hxi_connector.live_ingester.adapters.config.properties.Authorization;
 import org.alfresco.hxi_connector.live_ingester.domain.exception.LiveIngesterRuntimeException;
 
 @Service
@@ -62,11 +53,6 @@ import org.alfresco.hxi_connector.live_ingester.domain.exception.LiveIngesterRun
 @Slf4j
 public class AuthenticationService
 {
-    private static final String CLIENT_REGISTRATION_ID = "hyland-experience-auth";
-    private static final String APP_NAME_ATTRIBUTE_KEY = "applicationName";
-    private static final String SERVICE_USER_ATTRIBUTE_KEY = "serviceUser";
-    public static final String ENVIRONMENT_KEY_ATTRIBUTE_KEY = "hxAiEnvironmentKey";
-    private static final String ENVIRONMENT_KEY_HEADER = "hxai-environment";
     private static final int WAIT_FOR_PAUSE_TIME_MILLIS = 100;
 
     private final OAuth2ClientProperties oAuth2ClientProperties;
@@ -78,64 +64,22 @@ public class AuthenticationService
     @PostConstruct
     public void authenticationSchedule()
     {
-        if (isTokenUriNotBlank())
+        if (AuthSupport.isTokenUriNotBlank(oAuth2ClientProperties))
         {
             SecurityContext securityContext = SecurityContextHolder.getContext();
             DelegatingSecurityContextTaskScheduler delegatingTaskScheduler = new DelegatingSecurityContextTaskScheduler(taskScheduler, securityContext);
             Runnable authenticationTask = () -> {
                 waitFor(camelContext::isStarted);
-                authenticate();
+                String clientName = oAuth2ClientProperties.getRegistration().get(CLIENT_REGISTRATION_ID).getClientName();
+                String serviceUser = integrationProperties.hylandExperience().authorization().serviceUser();
+                String environmentKey = integrationProperties.hylandExperience().authorization().environmentKey();
+                OAuth2AuthenticationToken authenticationToken = AuthSupport.createOAuth2AuthenticationToken(clientName, serviceUser, environmentKey);
+                AuthSupport.authenticate(authenticationToken, authenticationManager);
             };
             delegatingTaskScheduler.scheduleWithFixedDelay(
                     authenticationTask,
                     Duration.ofMinutes(integrationProperties.hylandExperience().authentication().refreshDelayMinutes()));
         }
-    }
-
-    public void authenticate()
-    {
-        String clientName = oAuth2ClientProperties.getRegistration().get(CLIENT_REGISTRATION_ID).getClientName();
-        OAuth2AuthenticationToken authenticationToken = createOAuth2AuthenticationToken(clientName);
-        Authentication authentication = authenticationManager.authenticate(authenticationToken);
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-    }
-
-    public static void setAuthorizationToken(Exchange exchange)
-    {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (authentication instanceof OAuth2LoginAuthenticationToken authenticationToken)
-        {
-            OAuth2AccessToken accessToken = authenticationToken.getAccessToken();
-            Map<String, Object> principalAttributes = authenticationToken.getPrincipal().getAttributes();
-
-            String authorization = accessToken.getTokenType().getValue() + " " + accessToken.getTokenValue();
-            exchange.getIn().setHeaders(Map.of(
-                    AUTHORIZATION, authorization,
-                    ENVIRONMENT_KEY_HEADER, principalAttributes.get(ENVIRONMENT_KEY_ATTRIBUTE_KEY)));
-            log.debug("Authorization :: auth header added");
-        }
-        else
-        {
-            log.warn("Spring security context does not contain authentication principal of type " + OAuth2LoginAuthenticationToken.class.getSimpleName());
-        }
-    }
-
-    private OAuth2AuthenticationToken createOAuth2AuthenticationToken(String clientName)
-    {
-        Authorization authorizationProperties = integrationProperties.hylandExperience().authorization();
-        Map<String, Object> userAttributes = Map.of(
-                APP_NAME_ATTRIBUTE_KEY, clientName,
-                SERVICE_USER_ATTRIBUTE_KEY, authorizationProperties.serviceUser(),
-                ENVIRONMENT_KEY_ATTRIBUTE_KEY, authorizationProperties.environmentKey());
-        OAuth2UserAuthority oAuth2UserAuthority = new OAuth2UserAuthority(userAttributes);
-        OAuth2User oAuth2User = new DefaultOAuth2User(Set.of(oAuth2UserAuthority), userAttributes, APP_NAME_ATTRIBUTE_KEY);
-        return new OAuth2AuthenticationToken(oAuth2User, Set.of(oAuth2UserAuthority), CLIENT_REGISTRATION_ID);
-    }
-
-    private boolean isTokenUriNotBlank()
-    {
-        return oAuth2ClientProperties.getProvider().containsKey(CLIENT_REGISTRATION_ID)
-                && StringUtils.isNotBlank(oAuth2ClientProperties.getProvider().get(CLIENT_REGISTRATION_ID).getTokenUri());
     }
 
     private static void waitFor(Supplier<Boolean> supplier)
@@ -144,7 +88,7 @@ public class AuthenticationService
         {
             try
             {
-                Thread.sleep(WAIT_FOR_PAUSE_TIME_MILLIS);
+                TimeUnit.MILLISECONDS.sleep(WAIT_FOR_PAUSE_TIME_MILLIS);
             }
             catch (InterruptedException e)
             {
