+++ live-ingester/src/main/java/org/alfresco/hxi_connector/live_ingester/adapters/messaging/hx_insight/storage/connector/PreSignedUrlRequester.java
@@ -27,8 +27,8 @@ package org.alfresco.hxi_connector.live_ingester.adapters.messaging.hx_insight.s
 
 import static org.apache.camel.Exchange.HTTP_RESPONSE_CODE;
 
+import static org.alfresco.hxi_connector.common.adapters.auth.AuthSupport.setAuthorizationToken;
 import static org.alfresco.hxi_connector.common.util.ErrorUtils.UNEXPECTED_STATUS_CODE_MESSAGE;
-import static org.alfresco.hxi_connector.live_ingester.adapters.auth.AuthenticationService.setAuthorizationToken;
 
 import java.net.MalformedURLException;
 import java.net.URL;
